+++ UnitTests/UnitTests.csproj
@@ -19,6 +19,7 @@
     <TestProjectType>UnitTest</TestProjectType>
     <NuGetPackageImportStamp>
     </NuGetPackageImportStamp>
+    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -28,6 +29,7 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -36,6 +38,7 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="Microsoft.VisualStudio.TestPlatform.TestFramework, Version=14.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
@@ -48,13 +51,13 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ComboTests.cs" />
+    <Compile Include="src\ComboTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StealTests.cs" />
-    <Compile Include="Spell2Tests.cs" />
-    <Compile Include="CharacterTests.cs" />
-    <Compile Include="ChestTests.cs" />
-    <Compile Include="CircularBufferTests.cs" />
+    <Compile Include="src\StealTests.cs" />
+    <Compile Include="src\SpellTests.cs" />
+    <Compile Include="src\CharacterTests.cs" />
+    <Compile Include="src\ChestTests.cs" />
+    <Compile Include="src\CircularBufferTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
