+++ src/RNGHelper/ChestFutureRng.cs
