+++ src/RNGHelper/ChestRngHelper.cs
