+++ src/RNGHelper/Internal/IDeepClonable.cs
