+++ src/RNGHelper/Internal/IFutureRng.cs
