+++ src/RNGHelper/Internal/Steal.cs
