+++ src/RNGHelper/Internal/Utils.cs
@@ -4,7 +4,7 @@
     /// Class for holding methods used throughout
     /// the life of the program
     /// </summary>
-    public static class Utils
+    internal static class Utils
     {
         /// <summary>
         /// Convert a random number into its
