+++ src/RNGHelper/PlatformType.cs
